feat: path에 따른 컴포넌트 렌더링
