--- src/pages/Share/components/index.jsx
@@ -1,4 +1,0 @@
-import { Card } from "./Card";
-import { CardList } from "./CardList";
-
-export { Card, CardList };
